refactor: split workflow-version-step service into focused services
- Extract shared helpers into workflow-version-step-helpers service
- Create workflow-version-step-creation service for create/duplicate operations
- Create workflow-version-step-update service for update operations
- Create workflow-version-step-deletion service for deletion operations
- Reduce main service from 405 to 101 lines (75% reduction)
- Update tests to work with new service structure
