--- packages/twenty-server/src/modules/workflow/workflow-builder/workflow-version-step/constants/agent-naming.constants.ts
@@ -1,1 +1,0 @@
-export const MAX_AGENT_NAME_CONFLICT_ATTEMPTS = 100;
