+++ packages/twenty-server/src/modules/workflow/workflow-builder/workflow-version-step/workflow-version-step-helpers.workspace-service.ts
@@ -1,0 +1,65 @@
+import { Injectable } from '@nestjs/common';
+
+import { TwentyORMGlobalManager } from 'src/engine/twenty-orm/twenty-orm-global.manager';
+import { type WorkflowVersionWorkspaceEntity } from 'src/modules/workflow/common/standard-objects/workflow-version.workspace-entity';
+import { assertWorkflowVersionIsDraft } from 'src/modules/workflow/common/utils/assert-workflow-version-is-draft.util';
+import { WorkflowCommonWorkspaceService } from 'src/modules/workflow/common/workspace-services/workflow-common.workspace-service';
+import { type WorkflowAction } from 'src/modules/workflow/workflow-executor/workflow-actions/types/workflow-action.type';
+import { type WorkflowTrigger } from 'src/modules/workflow/common/types/workflow-trigger.type';
+
+@Injectable()
+export class WorkflowVersionStepHelpersWorkspaceService {
+  constructor(
+    private readonly twentyORMGlobalManager: TwentyORMGlobalManager,
+    private readonly workflowCommonWorkspaceService: WorkflowCommonWorkspaceService,
+  ) {}
+
+  async getValidatedDraftWorkflowVersion({
+    workflowVersionId,
+    workspaceId,
+  }: {
+    workflowVersionId: string;
+    workspaceId: string;
+  }): Promise<WorkflowVersionWorkspaceEntity> {
+    const workflowVersion =
+      await this.workflowCommonWorkspaceService.getWorkflowVersionOrFail({
+        workflowVersionId,
+        workspaceId,
+      });
+
+    assertWorkflowVersionIsDraft(workflowVersion);
+
+    return workflowVersion;
+  }
+
+  async updateWorkflowVersionStepsAndTrigger({
+    workspaceId,
+    workflowVersionId,
+    steps,
+    trigger,
+  }: {
+    workspaceId: string;
+    workflowVersionId: string;
+    steps?: WorkflowAction[];
+    trigger?: WorkflowTrigger;
+  }): Promise<void> {
+    const workflowVersionRepository =
+      await this.twentyORMGlobalManager.getRepositoryForWorkspace<WorkflowVersionWorkspaceEntity>(
+        workspaceId,
+        'workflowVersion',
+        { shouldBypassPermissionChecks: true },
+      );
+
+    const updateData: Partial<WorkflowVersionWorkspaceEntity> = {};
+
+    if (steps !== undefined) {
+      updateData.steps = steps;
+    }
+
+    if (trigger !== undefined) {
+      updateData.trigger = trigger;
+    }
+
+    await workflowVersionRepository.update(workflowVersionId, updateData);
+  }
+}
